Fixed a couple more save game issues

+++ src/domain/domain_game_data.h
@@ -153,7 +153,7 @@ class domain_game_data final : public QObject
 	Q_PROPERTY(QVariantList finished_journal_entries READ get_finished_journal_entries_qvariant_list NOTIFY journal_entries_changed)
 	Q_PROPERTY(int min_income READ get_min_income NOTIFY income_changed)
 	Q_PROPERTY(int max_income READ get_max_income NOTIFY income_changed)
-	Q_PROPERTY(int maintenance_cost READ get_maintenance_cost NOTIFY maintenance_cost_changed)
+	Q_PROPERTY(qint64 maintenance_cost READ get_maintenance_cost NOTIFY maintenance_cost_changed)
 	Q_PROPERTY(const metternich::site* visit_target_site READ get_visit_target_site WRITE set_visit_target_site NOTIFY visit_target_site_changed)
 
 public:
@@ -1214,8 +1214,8 @@ public:
 
 	int get_min_income() const;
 	int get_max_income() const;
-	int get_domain_maintenance_cost() const;
-	int get_maintenance_cost() const;
+	int64_t get_domain_maintenance_cost() const;
+	int64_t get_maintenance_cost() const;
 
 	bool has_flag(const flag *flag) const
 	{
